% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uzawa.r
\name{plot_results}
\alias{plot_results}
\title{Plot the results of the Uzawa algorithm}
\usage{
plot_results(x, y)
}
\arguments{
\item{x}{The solution vector x.}

\item{y}{The solution vector y.}
}
\description{
This function plots the solution vectors x and y.
}
\examples{
x <- c(1, 2)
y <- 3
plot_results(x, y)

}
