% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uzawa.r
\name{solve_system}
\alias{solve_system}
\title{Uzawa algorithm for saddle point problems}
\usage{
solve_system(...)
}
\arguments{
\item{A}{The matrix A in the saddle point problem Ax = b, Cx = d.}

\item{b}{The vector b in the saddle point problem Ax = b, Cx = d.}

\item{C}{The transpose of the matrix B in the saddle point problem Ax = b, Cx = d.}

\item{d}{The vector d in the saddle point problem Ax = b, Cx = d.}

\item{x0}{The initial guess for the solution vector x.}

\item{y0}{The initial guess for the solution vector y.}

\item{rho}{The penalty parameter.}

\item{alpha}{The step size.}

\item{max_iter}{The maximum number of iterations.}

\item{tol}{The tolerance for convergence.}
}
\value{
A list containing the solution vectors x and y.
}
\description{
This function implements the Uzawa algorithm for solving saddle point problems.
}
\examples{
A <- matrix(c(1, 2, 3, 4), nrow = 2)
b <- c(5, 6)
C <- matrix(c(1, 1), nrow = 1)
d <- 1
x0 <- c(0, 0)
y0 <- c(0)
rho <- 1
alpha <- 0.1
max_iter <- 1000
tol <- 1e-6
initialize_uzawa(A, b, C, d, x0, y0, rho, alpha, max_iter, tol)

}
